pyproject.toml
[tool.poetry]
name = "diploma"
version = "0.1.0"
description = ""
authors = ["danylo.filipenko <danylo.filipenko@gr8.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
supervision = "^0.19.0"


[tool.poetry.group.dev.dependencies]
absl-py = "2.0.0"
astunparse = "1.6.3"
cachetools = "5.3.2"
certifi = "2023.11.17"
charset-normalizer = "3.3.2"
contourpy = "1.2.1"
cvzone = "1.6.1"
cycler = "0.12.1"
defusedxml = "0.7.1"
filelock = "3.13.1"
flatbuffers = "23.5.26"
fonttools = "4.51.0"
fsspec = "2023.12.2"
gast = "0.5.4"
google-auth = "2.26.2"
google-auth-oauthlib = "1.2.0"
google-pasta = "0.2.0"
grpcio = "1.60.0"
h5py = "3.10.0"
idna = "3.6"
jinja2 = "3.1.3"
keras = "2.15.0"
kiwisolver = "1.4.5"
libclang = "16.0.6"
markdown = "3.5.2"
markupsafe = "2.1.3"
matplotlib = "3.8.4"
ml-dtypes = "0.2.0"
mpmath = "1.3.0"
networkx = "3.2.1"
numpy = "1.26.4"
oauthlib = "3.2.2"
opencv-python = "4.9.0.80"
opencv-python-headless = "4.9.0.80"
opt-einsum = "3.3.0"
packaging = "24.0"
pandas = "2.1.4"
pillow = "10.3.0"
protobuf = "4.23.4"
psutil = "5.9.7"
py-cpuinfo = "9.0.0"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pyparsing = "3.1.2"
python-dateutil = "2.9.0.post0"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
rsa = "4.9"
scipy = "1.13.0"
seaborn = "0.13.1"
six = "1.16.0"
supervision = "0.19.0"
sympy = "1.12"
tensorboard = "2.15.1"
tensorboard-data-server = "0.7.2"
tensorflow-estimator = "2.15.0"
tensorflow-io-gcs-filesystem = "0.34.0"
tensorflow-macos = "2.15.0"
termcolor = "2.4.0"
thop = "0.1.1.post2209072238"
tk = "0.1.0"
torch = "2.1.2"
torchvision = "0.16.2"
tqdm = "4.66.1"
typing-extensions = "4.9.0"
tzdata = "2023.4"
ultralytics = "8.1.2"
urllib3 = "2.1.0"
werkzeug = "3.0.1"
wrapt = "1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
